Submission for Kaggle Spaceship Titanic Prediction Comp
Submission Date: 1/6/25

Practiced EDA visualization on each column individually to determine the best method for handling missing values in that column. Expanded on feature engineering process with the creation of multiple new features using existing data in the dataset. Built a CatBoostClassifier as the prediction model.

Score: 0.80593
Place (as of 1/6/25): 209 / 2340